@@ -2,7 +2,8 @@
   "name": "bbzwgucci-store",
   "version": "0.0.0",
   "scripts": {
-    "start": "node backend/server",
+    "start": "npm run build && npm run start-backend",
+    "start-backend": "node backend/server",
     "build": "npm run compileBackend && ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -10,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/animations": "9.1.13",
+    "@angular/common": "9.1.13",
+    "@angular/compiler": "9.1.13",
+    "@angular/compiler-cli": "9.1.13",
+    "@angular/core": "9.1.13",
+    "@angular/forms": "9.1.13",
+    "@angular/platform-browser": "9.1.13",
+    "@angular/platform-browser-dynamic": "9.1.13",
+    "@angular/router": "9.1.13",
     "@ng-bootstrap/ng-bootstrap": "5.1.5",
     "@sweetalert2/ngx-sweetalert2": "7.2.0",
     "bootstrap": "4.4.1",
@@ -28,17 +29,17 @@
     "npm": "6.13.6",
     "rxjs": "6.5.4",
     "sweetalert2": "9.6.0",
-    "tslib": "1.10.0",
-    "zone.js": "0.10.2"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.23",
-    "@angular/cli": "8.3.23",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular/cli": "9.1.15",
+    "@angular/language-service": "9.1.13",
     "@types/jasmine": "3.5.0",
     "@types/jasminewd2": "2.0.8",
-    "@types/node": "13.1.7",
-    "codelyzer": "5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "3.5.0",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "4.4.1",
@@ -49,6 +50,6 @@
     "protractor": "5.4.2",
     "ts-node": "8.6.2",
     "tslint": "5.20.1",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   }
 }
